requirements.txt
python-dotenv==0.19.0
slack-bolt==1.7.0
slack-sdk==3.8.0
